'finish-booklist-element-messagebox'

+++ src/components/NoteBookList.vue
@@ -25,8 +25,7 @@
     import Auth from '@/apis/auth'
     import NoteBooks from '@/apis/notebook'
     // import notebook from "../apis/notebook";
-    import {friendlyDate} from "../helpers/util";
-    window.notebook = NoteBooks
+    import {friendlyDate} from "../helpers/util"
   export default {
     data () {
       return {
@@ -50,37 +49,57 @@
     },
     methods: {
       onCreate () {
-        console.log('on create')
-        let title = window.prompt('创建笔记本')
-        if(title.trim()===''){
-          alert('笔记本不能为空')
-          return
-        }
-        NoteBooks.addNoteBook({ title})
-          .then(res=>{
-            console.log(res.data)
-            alert(res.msg)
-            res.data.friendlycreatedAt = friendlyDate(res.data.createdAt)
-            this.notebooks.unshift(res.data)
-          })
+        this.$prompt('请输入新笔记标题', '创建笔记本', {
+          confirmButtonText: '确定',
+          cancelButtonText: '取消',
+          inputPattern: /^.{1,30}$/,
+          inputErrorMessage: '笔记本标题不能为空，且不超30个字'
+        }).then(({ value }) => {
+          return  NoteBooks.addNoteBook({ title: value})
+        }).then(res=>{
+          res.data.friendlycreatedAt = friendlyDate(res.data.createdAt)
+          this.notebooks.unshift(res.data)
+          this.$message({
+            type: 'success',
+            message: res.msg
+          });
+        }).catch(()=>{})
+
       },
       onEdit (notebook) {
-        console.log('on EDIT')
-        let title = window.prompt('修改标题', notebook.title)
-        NoteBooks.updateNotebook(notebook.id,{ title })
-          .then(res=>{
-            console.log(res.data)
-            alert('修改成功')
-            notebook.title = title
+        let notebookTitle = ''
+        this.$prompt('请修改笔记本', '修改笔记本', {
+          confirmButtonText: '确定',
+          cancelButtonText: '取消',
+          inputPattern: /^.{1,30}$/,
+          inputValue: notebook.title,
+          inputErrorMessage: '笔记本标题不能为空，且不超30个字'
+        }).then(({ value }) => {
+          notebookTitle = value
+          return  NoteBooks.updateNotebook(notebook.id,{ title: value })
+        }).then(res=>{
+          notebook.title = notebookTitle
+          this.$message({
+            type: 'success',
+            message: res.msg
           })
+        }).catch(()=>{})
+
       },
       onDelete (notebook) {
-        console.log('on delete')
-        NoteBooks.deleteNotebook(notebook.id)
-          .then(res=>{
-            alert(res.msg)
-            this.notebooks.splice(this.notebooks.indexOf(notebook), 1)
-          })
+        this.$confirm('确定要删除笔记本?', '删除笔记本', {
+          confirmButtonText: '确定',
+          cancelButtonText: '取消',
+          type: 'warning'
+        }).then(() => {
+          return NoteBooks.deleteNotebook(notebook.id)
+        }).then(()=>{
+          this.$message({
+            type: 'success',
+            message: '删除成功!'
+          });
+          this.notebooks.splice(this.notebooks.indexOf(notebook), 1)
+      }).catch(()=>{})
       }
     }
   }
